Account for WriteLine impedance mismatch between PowerShell and the MSBuild log
The PowerShell host needs to expose the ability to to write text without indicating a new
line. On the other hand, the MSBuild log doesn't support this notion. As a result,
we have to buffer messages that are passed to the PowerShell host until we receive
a new-line character (yuck).